Consolidate custom Auth call ugliness to logout
With #47 in place, the only place we have in the portal code now that
needs access to Auth configuration (e.g. logout URI, client ID/secret)
for making custom-coded calls is the logout route.

Given this, we can consolidate all the custom code for dealing with
token revocation and redirecting to the logout URI to this one route.